+++ internal/checks/promql_comparison_test.go
@@ -11,31 +11,31 @@ func TestComparisonCheck(t *testing.T) {
 		{
 			description: "ignores recording rules",
 			content:     "- record: foo\n  expr: up == 0\n",
-			checker:     checks.NewComparisonCheck(checks.Bug),
+			checker:     checks.NewComparisonCheck(),
 		},
 		{
 			description: "ignores rules with syntax errors",
 			content:     "- alert: Foo Is Down\n  expr: sum(\n",
-			checker:     checks.NewComparisonCheck(checks.Bug),
+			checker:     checks.NewComparisonCheck(),
 		},
 		{
 			description: "alert expr with > condition",
 			content:     "- alert: Foo Is Down\n  for: 10m\n  expr: up{job=\"foo\"} > 0\n",
-			checker:     checks.NewComparisonCheck(checks.Bug),
+			checker:     checks.NewComparisonCheck(),
 		},
 		{
 			description: "alert expr with >= condition",
 			content:     "- alert: Foo Is Down\n  for: 10m\n  expr: up{job=\"foo\"} >= 1\n",
-			checker:     checks.NewComparisonCheck(checks.Bug)},
+			checker:     checks.NewComparisonCheck()},
 		{
 			description: "alert expr with == condition",
 			content:     "- alert: Foo Is Down\n  for: 10m\n  expr: up{job=\"foo\"} == 1\n",
-			checker:     checks.NewComparisonCheck(checks.Bug),
+			checker:     checks.NewComparisonCheck(),
 		},
 		{
 			description: "alert expr without any condition",
 			content:     "- alert: Foo Is Down\n  expr: up{job=\"foo\"}\n",
-			checker:     checks.NewComparisonCheck(checks.Warning),
+			checker:     checks.NewComparisonCheck(),
 			problems: []checks.Problem{
 				{
 					Fragment: `up{job="foo"}`,
@@ -54,7 +54,7 @@ func TestComparisonCheck(t *testing.T) {
         AND ON (instance)
         (rate(node_netstat_Udp_RcvbufErrors[5m])+rate(node_netstat_Udp6_RcvbufErrors[5m])) > 200
 `,
-			checker: checks.NewComparisonCheck(checks.Warning),
+			checker: checks.NewComparisonCheck(),
 		},
 		{
 			description: "deep level without comparison",
@@ -64,7 +64,7 @@ func TestComparisonCheck(t *testing.T) {
         AND ON (instance)
         rate(node_netstat_Udp_RcvbufErrors[5m])+rate(node_netstat_Udp6_RcvbufErrors[5m])
 `,
-			checker: checks.NewComparisonCheck(checks.Warning),
+			checker: checks.NewComparisonCheck(),
 			problems: []checks.Problem{
 				{
 					Fragment: `quantile_over_time(0.7,(irate(udp_packets_drops[2m]))[10m:2m]) AND ON (instance) rate(node_netstat_Udp_RcvbufErrors[5m])+rate(node_netstat_Udp6_RcvbufErrors[5m])`,
@@ -78,31 +78,31 @@ func TestComparisonCheck(t *testing.T) {
 		{
 			description: "alert unless condition",
 			content:     "- alert: Foo Is Down\n  for: 10m\n  expr: foo unless bar\n",
-			checker:     checks.NewComparisonCheck(checks.Bug),
+			checker:     checks.NewComparisonCheck(),
 		},
 		{
 			description: "alert expr with bool",
 			content:     "- alert: Error rate is high\n  expr: rate(error_count[5m]) > bool 5\n",
-			checker:     checks.NewComparisonCheck(checks.Warning),
+			checker:     checks.NewComparisonCheck(),
 			problems: []checks.Problem{
 				{
 					Fragment: "rate(error_count[5m]) > bool 5",
 					Lines:    []int{2},
 					Reporter: "promql/comparison",
 					Text:     "alert query uses bool modifier for comparison, this means it will always return a result and the alert will always fire",
-					Severity: checks.Warning,
+					Severity: checks.Bug,
 				},
 			},
 		},
 		{
 			description: "alert expr with bool and condition",
 			content:     "- alert: Error rate is high\n  expr: rate(error_count[5m]) > bool 5 == 1\n",
-			checker:     checks.NewComparisonCheck(checks.Warning),
+			checker:     checks.NewComparisonCheck(),
 		},
 		{
 			description: "alert on absent",
 			content:     "- alert: Foo Is Missing\n  expr: absent(foo)\n",
-			checker:     checks.NewComparisonCheck(checks.Bug),
+			checker:     checks.NewComparisonCheck(),
 		},
 	}