+++ internal/checks/promql_vectormatching_test.go
@@ -161,56 +161,56 @@ func TestVectorMatchingCheck(t *testing.T) {
 		{
 			description: "ignores rules with syntax errors",
 			content:     "- record: foo\n  expr: sum(foo) without(\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "one to one matching",
 			content:     "- record: foo\n  expr: foo_with_notfound / bar\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 			problems: []checks.Problem{
 				{
 					Fragment: "foo_with_notfound / bar",
 					Lines:    []int{2},
 					Reporter: "promql/vector_matching",
 					Text:     `both sides of the query have different labels: [instance job notfound] != [instance job]`,
-					Severity: checks.Warning,
+					Severity: checks.Bug,
 				},
 			},
 		},
 		{
 			description: "ignore left query errors",
 			content:     "- record: foo\n  expr: xxx / foo\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "ignore righ query errors",
 			content:     "- record: foo\n  expr: foo / xxx\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "ignore missing left series",
 			content:     "- record: foo\n  expr: missing / foo\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "ignore missing or vector",
 			content:     "- record: foo\n  expr: sum(missing or vector(0))\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "ignore present or vector",
 			content:     "- record: foo\n  expr: sum(foo or vector(0))\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "ignore missing right series",
 			content:     "- record: foo\n  expr: foo / missing\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "ignore with mismatched series",
 			content:     "- record: foo\n  expr: foo / ignoring(xxx) app_registry\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Bug),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 			problems: []checks.Problem{
 				{
 					Fragment: "foo / ignoring(xxx) app_registry",
@@ -224,81 +224,81 @@ func TestVectorMatchingCheck(t *testing.T) {
 		{
 			description: "one to one matching with on() - both missing",
 			content:     "- record: foo\n  expr: foo / on(notfound) bar\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 			problems: []checks.Problem{
 				{
 					Fragment: "foo / on(notfound) bar",
 					Lines:    []int{2},
 					Reporter: "promql/vector_matching",
 					Text:     `using on("notfound") won't produce any results because both sides of the query don't have this label`,
-					Severity: checks.Warning,
+					Severity: checks.Bug,
 				},
 			},
 		},
 		{
 			description: "one to one matching with ignoring() - both missing",
 			content:     "- record: foo\n  expr: foo / ignoring(notfound) foo\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "one to one matching with ignoring() - both present",
 			content:     "- record: foo\n  expr: foo_with_notfound / ignoring(notfound) foo_with_notfound\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "one to one matching with ignoring() - left missing",
 			content:     "- record: foo\n  expr: foo / ignoring(notfound) foo_with_notfound\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "one to one matching with ignoring() - right missing",
 			content:     "- record: foo\n  expr: foo_with_notfound / ignoring(notfound) foo\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "one to one matching with ignoring() - mismatch",
 			content:     "- record: foo\n  expr: foo_with_notfound / ignoring(notfound) bar\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 		{
 			description: "one to one matching with on() - left missing",
 			content:     "- record: foo\n  expr: foo / on(notfound) bar_with_notfound\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 			problems: []checks.Problem{
 				{
 					Fragment: "foo / on(notfound) bar_with_notfound",
 					Lines:    []int{2},
 					Reporter: "promql/vector_matching",
 					Text:     `using on("notfound") won't produce any results because left hand side of the query doesn't have this label: "foo"`,
-					Severity: checks.Warning,
+					Severity: checks.Bug,
 				},
 			},
 		},
 		{
 			description: "one to one matching with on() - right missing",
 			content:     "- record: foo\n  expr: foo_with_notfound / on(notfound) bar\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 			problems: []checks.Problem{
 				{
 					Fragment: "foo_with_notfound / on(notfound) bar",
 					Lines:    []int{2},
 					Reporter: "promql/vector_matching",
 					Text:     `using on("notfound") won't produce any results because right hand side of the query doesn't have this label: "bar"`,
-					Severity: checks.Warning,
+					Severity: checks.Bug,
 				},
 			},
 		},
 		{
 			description: "nested query",
 			content:     "- alert: foo\n  expr: (memory_bytes / ignoring(job) (memory_limit > 0)) * on(app_name) group_left(a,b,c) app_registry\n",
-			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker:     checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 			problems: []checks.Problem{
 				{
 					Fragment: "(memory_bytes / ignoring(job) (memory_limit > 0)) * on(app_name) group_left(a,b,c) app_registry",
 					Lines:    []int{2},
 					Reporter: "promql/vector_matching",
 					Text:     `using on("app_name") won't produce any results because left hand side of the query doesn't have this label: "(memory_bytes / ignoring(job) (memory_limit > 0))"`,
-					Severity: checks.Warning,
+					Severity: checks.Bug,
 				},
 			},
 		},
@@ -308,7 +308,7 @@ func TestVectorMatchingCheck(t *testing.T) {
 - record: foo
   expr: '{__name__="foo_with_notfound"} / ignoring(notfound) foo_with_notfound'
 `,
-			checker: checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1, checks.Warning),
+			checker: checks.NewVectorMatchingCheck("prom", srv.URL, time.Second*1),
 		},
 	}
 	runTests(t, testCases)