+++ internal/checks/query_series_test.go
@@ -108,12 +108,12 @@ func TestSeriesCheck(t *testing.T) {
 		{
 			description: "ignores rules with syntax errors",
 			content:     "- record: foo\n  expr: sum(foo) without(\n",
-			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5, checks.Warning),
+			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5),
 		},
 		{
 			description: "bad response",
 			content:     "- record: foo\n  expr: sum(foo)\n",
-			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5, checks.Warning),
+			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5),
 			problems: []checks.Problem{
 				{
 					Fragment: "foo",
@@ -127,7 +127,7 @@ func TestSeriesCheck(t *testing.T) {
 		{
 			description: "simple query",
 			content:     "- record: foo\n  expr: sum(notfound)\n",
-			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5, checks.Warning),
+			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5),
 			problems: []checks.Problem{
 				{
 					Fragment: "notfound",
@@ -141,7 +141,7 @@ func TestSeriesCheck(t *testing.T) {
 		{
 			description: "complex query",
 			content:     "- record: foo\n  expr: sum(found_7 * on (job) sum(sum(notfound))) / found_7\n",
-			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5, checks.Warning),
+			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5),
 			problems: []checks.Problem{
 				{
 					Fragment: "notfound",
@@ -155,14 +155,14 @@ func TestSeriesCheck(t *testing.T) {
 		{
 			description: "complex query / bug",
 			content:     "- record: foo\n  expr: sum(found_7 * on (job) sum(sum(notfound))) / found_7\n",
-			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5, checks.Bug),
+			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5),
 			problems: []checks.Problem{
 				{
 					Fragment: "notfound",
 					Lines:    []int{2},
 					Reporter: "query/series",
 					Text:     "query using prom completed without any results for notfound",
-					Severity: checks.Bug,
+					Severity: checks.Warning,
 				},
 			},
 		},
@@ -189,17 +189,17 @@ func TestSeriesCheck(t *testing.T) {
     )
   for: 5m
 `,
-			checker: checks.NewSeriesCheck("prom", srv.URL, time.Second*5, checks.Bug),
+			checker: checks.NewSeriesCheck("prom", srv.URL, time.Second*5),
 		},
 		{
 			description: "offset",
 			content:     "- record: foo\n  expr: node_filesystem_readonly{mountpoint!=\"\"} offset 5m\n",
-			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5, checks.Bug),
+			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5),
 		},
 		{
 			description: "series found, label missing",
 			content:     "- record: foo\n  expr: found{job=\"notfound\"}\n",
-			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5, checks.Warning),
+			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5),
 			problems: []checks.Problem{
 				{
 					Fragment: `found{job="notfound"}`,
@@ -213,7 +213,7 @@ func TestSeriesCheck(t *testing.T) {
 		{
 			description: "series missing, label missing",
 			content:     "- record: foo\n  expr: notfound{job=\"notfound\"}\n",
-			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5, checks.Warning),
+			checker:     checks.NewSeriesCheck("prom", srv.URL, time.Second*5),
 			problems: []checks.Problem{
 				{
 					Fragment: "notfound",
@@ -230,7 +230,7 @@ func TestSeriesCheck(t *testing.T) {
 - record: foo
   expr: '{__name__="notfound", job="bar"}'
 `,
-			checker: checks.NewSeriesCheck("prom", srv.URL, time.Second*5, checks.Warning),
+			checker: checks.NewSeriesCheck("prom", srv.URL, time.Second*5),
 			problems: []checks.Problem{
 				{
 					Fragment: `{__name__="notfound",job="bar"}`,