+++ internal/config/rate.go
@@ -1,8 +1,0 @@
-package config
-
-type RateSettings struct {
-}
-
-func (rs RateSettings) validate() error {
-	return nil
-}